+++ backendservice/src/main/java/com/google/gpay/virtualqueue/backendservice/repository/VirtualQueueRepository.java
@@ -18,7 +18,6 @@ package com.google.gpay.virtualqueue.backendservice.repository;
 
 import java.util.List;
 import java.util.UUID;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import com.google.gpay.virtualqueue.backendservice.model.Shop;
 import com.google.gpay.virtualqueue.backendservice.model.Token;
@@ -47,7 +46,7 @@ public interface VirtualQueueRepository {
 
     public Token getTokenByTokenId(UUID tokenId);
 
-    public AtomicInteger getCustomersAheadByTokenId(UUID tokenId);
+    public long getCustomersAheadByTokenId(UUID tokenId);
 
     public String getShopNameByTokenId(UUID tokenId);
 }
